requirements.txt
web3>=6.17.0
tls-client>=1.0.1
loguru>=0.7.2
Faker>=25.0.1
requests>=2.31.0
termcolor>=2.4.0
questionary>=2.0.1
python-dotenv>=1.0.1
aiohttp>=3.9.5
web3>=6.19.0
aiohttp-socks>=0.8.4
schedule
python-jose==3.3.0
